Some major changes. More modular structure: separated Anchor and AnchorManager, created the new headers for Math and Log Utils. Fixed the anchor updating algorithm was incorrect as the device pose was being calculate for the history change timestamp and it has to be calculated for each anchor timestamp specifically.
