+++ android_webview/test/shell/tango/jni/ndkbuild.sh
@@ -37,10 +37,20 @@ EXIT_CODE=$?
 if [ $EXIT_CODE -ne 0 ]; then 
 	exit $EXIT_CODE
 fi
-cp *.h ../../../../../third_party/tango/libtango_chromium
+cp TangoHandler.h ../../../../../third_party/tango/libtango_chromium
 EXIT_CODE=$?
 if [ $EXIT_CODE -ne 0 ]; then 
 	exit $EXIT_CODE
+fi
+cp Anchor.h ../../../../../third_party/tango/libtango_chromium
+EXIT_CODE=$?
+if [ $EXIT_CODE -ne 0 ]; then 
+  exit $EXIT_CODE
+fi
+cp AnchorManager.h ../../../../../third_party/tango/libtango_chromium
+EXIT_CODE=$?
+if [ $EXIT_CODE -ne 0 ]; then 
+  exit $EXIT_CODE
 fi
 echo "Rebuilt!"
